+++ SSDProject/TestShell/TestShell.vcxproj.filters
@@ -59,8 +59,5 @@
     <ClInclude Include="Command.h">
       <Filter>소스 파일</Filter>
     </ClInclude>
-    <ClInclude Include="Logger.h">
-      <Filter>헤더 파일</Filter>
-    </ClInclude>
   </ItemGroup>
 </Project>
